dlt_source_personio/model/v1/__init__.py
from pydantic import BaseModel


class MyBaseModel(BaseModel):
    pass


class MyAuthBaseModel(MyBaseModel):
    pass


class MyPersonnelBaseModel(MyBaseModel):
    pass


class MyRecruitingBaseModel(MyBaseModel):
    pass
