pyproject.toml
[project]
name = "dlt-source-personio"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dlt>=1.8.0",
    "pydantic-extra-types>=2.10.3",
]
dynamic = ["authors", "classifiers", "version", "description"]

[project.optional-dependencies]
show = [
    "dlt[duckdb]>=1.8.0",
    "streamlit>=1.41.1",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "datamodel-code-generator>=0.28.3",
    "pytest>=8.3.4",
]

[tool.black]
include = '.*py$'
